+++ tests/data/errors/900_bad_token.sql
@@ -1,3 +1,1 @@
-select ?
-from my_table
-where no_question_mark is true
+select $